@@ -38,82 +38,63 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "14.2.0",
-		"@angular/common": "14.2.0",
-		"@angular/compiler": "14.2.0",
-		"@angular/core": "14.2.0",
-		"@angular/forms": "14.2.0",
-		"@angular/platform-browser": "14.2.0",
-		"@angular/platform-browser-dynamic": "14.2.0",
-		"@angular/router": "14.2.0",
+		"@angular/animations": "16.1.3",
+		"@angular/common": "16.1.3",
+		"@angular/compiler": "16.1.3",
+		"@angular/core": "16.1.3",
+		"@angular/forms": "16.1.3",
+		"@angular/platform-browser": "16.1.3",
+		"@angular/platform-browser-dynamic": "16.1.3",
+		"@angular/router": "16.1.3",
+		"@nx/angular": "16.4.0",
+		"rxjs": "~7.8.0",
+		"tslib": "^2.3.0",
+		"zone.js": "~0.13.0",
 		"@ngneat/lib": "^3.0.1",
-		"@nrwl/angular": "14.6.3",
-		"angular-svg-icon": "^13.0.0",
-		"rxjs": "^6.5.3 || ^7.4.0",
-		"tslib": "^2.0.0",
-		"zone.js": "0.11.4"
+		"angular-svg-icon": "^16.0.0"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "14.2.1",
-		"@angular-eslint/eslint-plugin": "14.0.3",
-		"@angular-eslint/eslint-plugin-template": "14.0.3",
-		"@angular/cli": "~14.2.0",
-		"@angular/compiler-cli": "14.2.0",
-		"@angular/language-service": "14.2.0",
-		"@commitlint/cli": "^16.1.0",
-		"@commitlint/config-angular": "^8.1.0",
-		"@commitlint/config-conventional": "^8.1.0",
-		"@ngneat/spectator": "latest",
-		"@nrwl/cli": "14.6.3",
-		"@nrwl/cypress": "14.6.3",
-		"@nrwl/eslint-plugin-nx": "14.6.3",
-		"@nrwl/jest": "14.6.3",
-		"@nrwl/nx-cloud": "14.6.1",
-		"@nrwl/tao": "14.6.3",
-		"@nrwl/workspace": "14.6.3",
-		"@types/jest": "28.1.8",
-		"@types/node": "14.14.33",
-		"@typescript-eslint/eslint-plugin": "5.33.1",
-		"@typescript-eslint/parser": "5.33.1",
-		"all-contributors-cli": "^6.8.1",
-		"codelyzer": "^0.0.28",
-		"cross-env": "^5.2.0",
-		"cypress": "^8.3.0",
-		"dotenv": "10.0.0",
-		"eslint": "7.22.0",
+		"@angular-devkit/build-angular": "16.1.3",
+		"@angular-devkit/core": "16.1.3",
+		"@angular-devkit/schematics": "16.1.3",
+		"@angular-eslint/eslint-plugin": "~16.0.0",
+		"@angular-eslint/eslint-plugin-template": "~16.0.0",
+		"@angular-eslint/template-parser": "~16.0.0",
+		"@angular/cli": "~16.1.0",
+		"@angular/compiler-cli": "16.1.3",
+		"@angular/language-service": "16.1.3",
+		"@nx/cypress": "16.4.0",
+		"@nx/eslint-plugin": "16.4.0",
+		"@nx/jest": "16.4.0",
+		"@nx/js": "16.4.0",
+		"@nx/linter": "16.4.0",
+		"@nx/web": "16.4.0",
+		"@nx/workspace": "16.4.0",
+		"@schematics/angular": "16.1.3",
+		"@types/jest": "^29.4.0",
+		"@types/node": "16.11.7",
+		"@typescript-eslint/eslint-plugin": "^5.58.0",
+		"@typescript-eslint/parser": "^5.58.0",
+		"autoprefixer": "^10.4.0",
+		"cypress": "^12.11.0",
+		"eslint": "~8.15.0",
 		"eslint-config-prettier": "8.1.0",
 		"eslint-plugin-cypress": "^2.10.3",
+		"jest": "^29.4.1",
+		"jest-environment-jsdom": "^29.4.1",
+		"jest-preset-angular": "~13.1.0",
+		"ng-packagr": "16.1.0",
+		"nx": "16.4.0",
+		"nx-cloud": "16.0.5",
 		"git-cz": "^3.2.1",
-		"husky": "^3.0.1",
-		"jest": "28.1.3",
-		"jest-preset-angular": "12.2.2",
-		"lint-staged": "^12.3.2",
-		"ng-packagr": "14.2.0",
 		"postcss": "^8.3.9",
 		"postcss-import": "14.1.0",
 		"postcss-preset-env": "7.5.0",
 		"postcss-url": "10.1.3",
 		"prettier": "2.7.1",
 		"standard-version": "^9.3.2",
-		"ts-jest": "28.0.8",
+		"ts-jest": "29.1.0",
 		"ts-node": "10.9.1",
-		"typescript": "4.7.4"
-	},
-	"config": {
-		"commitizen": {
-			"path": "cz-conventional-changelog"
-		}
-	},
-	"lint-staged": {
-		"*.{js,json,css,scss,ts,html,component.html}": [
-			"prettier --write",
-			"git add"
-		]
-	},
-	"husky": {
-		"hooks": {
-			"pre-commit": "npm run hooks:pre-commit && lint-staged",
-			"pre-push": "npm run test:lib:headless"
-		}
+		"typescript": "5.1.6"
 	}
 }
