@@ -2,17 +2,16 @@ module.exports = {
 	displayName: 'angular-svg-icon-preloader',
 	preset: '../../jest.preset.js',
 	setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-	globals: {
-		'ts-jest': {
-			tsconfig: '<rootDir>/tsconfig.spec.json',
-			stringifyContentPathRegex: '\\.(html|svg)$',
-		},
-	},
-	coverageDirectory: '../../coverage/libs/angular-svg-icon-preloader',
 	transform: {
-		'^.+.(ts|mjs|js|html)$': 'jest-preset-angular',
+		'^.+\\.(ts|mjs|js|html)$': [
+			'jest-preset-angular',
+			{
+				tsconfig: '<rootDir>/tsconfig.spec.json',
+				stringifyContentPathRegex: '\\.(html|svg)$',
+			},
+		],
 	},
-	transformIgnorePatterns: ['node_modules/(?!.*.mjs$)'],
+	transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
 	snapshotSerializers: [
 		'jest-preset-angular/build/serializers/no-ng-attributes',
 		'jest-preset-angular/build/serializers/ng-snapshot',
